WIP (back): add accounts/:id/list_transactions
Function isn't hooked up to endpoint yet, and moreover, there seem to be weird pockets of non-variance in the descriptions of transactions, most likely related to the math of how I'm generating the variance. Going to try to use UUIDs to get more variance between transactions on the same day.